ProjectWarrier/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
openai
fastapi-cors
python-dotenv
greenlet
gTTS
pygame
requests
pvporcupine
pyaudio
webrtcvad-wheels
sounddevice
openai-whisper